qemu: address: Enable auto addressing multifunction cards
For existing domains using the primary function alone of a multifunction
card, the card is still treated as a multifunction card. This is done
to prevent hotplug of other functions when the primary function is
already hotplugged.

If the secondary functions are part of the xml without the primary
function being part of the xml, this has never been supported. So,
Libvirt doesn't consider this either as a multifunction card.

Since we're now checking PCI headers via
virHostdevIsPCIMultifunctionDevice(), changes in virpcitestdata
files were required to allow the test suit to recognize the
0005:90:01.N test device as multifunction.

Signed-off-by: Shivaprasad G Bhat <sbhat@linux.vnet.ibm.com>
Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>