Introduce virDomainDeviceDefParseXMLMany
virDomainDeviceDefParseXMLMany will be used to parse '<devices>'
XML elements that will be used to supply multiple hostdevs
for hotplug/unplug. The idea is to parse each dev individually,
by calling the same code that already handles it today. The
result is returned in a new virDomainDeviceDefListPtr pointer
type, which is an array of device definitions.

This idea of calling existing functions to handles a single
device definition in a loop is the central idea of how PCI
multifunction hotplug and hotunplug is implemented in the
next patches. Roughly speaking, we'll use virDomainDeviceDefListPtr
lists to call (almost) the same hotplug/unplug mechanics we have
today for each hostdev in the list.

Signed-off-by: Shivaprasad G Bhat <sbhat@linux.vnet.ibm.com>
Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>