qemu: hotplug: Implement multifunction device hotplug
The support to allow multifunction device hotplug consists
of using virDomainDeviceDefListPtr lists instead of single
virDomainDeviceDefPtr device definitions in both
qemuDomainAttachDeviceConfig() and qemuDomainAttachDeviceLive().
In AttachDeviceConfig() the same existing single device code
is ran for all devs in the list.

AttachDeviceLive() will verify if the list has size = 1 (meaning it
is a single device attach) and redirect the code to the regular single
dev attach procedure. If list size > 1, we'll execute a new specialized
function called qemuDomainAttachMultifunctionDevice(), which is
prepared to handle the nuances of this type of hotplug.

The changes in both AttachDeviceLive() and AttachDeviceConfig()
implied in changes to be made in qemuDomainAttachDeviceLiveAndConfig(),
which is now handling a device list instead of a single device.

All other added functions, most notably
qemuDomainPCIMultifunctionHostdevEnsurePCIAddresses(), are used
inside qemuDomainAttachMultifunctionDevice().

No changes in regular device hotplug mechanics were made.

Signed-off-by: Shivaprasad G Bhat <sbhat@linux.vnet.ibm.com>
Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>