qemu_hotplug.c: introduce qemuDomainWaitForMultipleDeviceRemoval
This new function is in fact the existing qemuDomainWaitForDeviceRemoval
with an extra parameter to handle the case of multifunction device
unplug. In a multifunction unplug, QEMU will unplug all the functions
in a single detach operation, sequentially, and this impacts the
amount of timeout we're supposed to wait. The previous existing
timeout wasn't enough to handle the multi-function detach case,
timing out the detach operation needlesly.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>