qemu_driver: use VIR_AUTOUNREF() with virQEMUDriverConfigPtr 2/3
virQEMUDriverConfigPtr can be auto-unref for the great majority
of the uses made in qemu_driver, sparing us a virObjectUnref()
call and sometimes a whole 'cleanup' label.

This patch changes virQEMUDriverConfigPtr declarations to
use VIR_AUTOUNREF(). 'cleanup' labels were deleted when
applicable.

Since there are a lot of references to change, let's do it in
3 steps. This is step 2.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>