qemu_domain: validate virtio in network validate functions
Add virtio validation in qemuDomainDeviceDefValidateNetwork(),
replicating what's being checked in qemuBuildNicDevStr() via
qemuBuildVirtioDevStr(), for VIR_DOMAIN_DEVICE_NET.

qemuBuildNicDevStr() is also called in qemu_hotplug.c, in
qemuDomainAttachNetDevice(). To keep this function covered with the
same validation capabilities after we remove the redundant validations
from qemuBuildVirtioDevStr() in a later patch, virtio validation must
be added.

There is already a function that validates NET devices in this
function, qemuDomainValidateActualNetDef(), that differs from the
validation function used in domain definition time because it relies
on information attained only in runtime. Let's use it to call the
virtio validation function directly.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>