qemu_domain: remove virtio verification from qemuBuildVirtioDevStr
Previous patches added qemuDomainDefValidateVirtioDev(), a
function inside qemu_domain.c that does the virtio verification
that qemuBuildVirtioDevStr() from qemu_command.c is doing.
The goal is to centralize this verification in domain
definition time instead of domain start time.

qemuBuildVirtioDevStr() is used in a lot of places inside
qemu_command.c, and previous patches needed to cover up
all the verifications that were going to be needed after
we remove the virtio check in qemuBuildVirtioDevStr().

We're able to remove the virtio check code from
qemuBuildVirtioDevStr() without the fear of leaving use cases
uncovered. The helper function qemuDomainGetVirtioTransitional(),
used by qemuDomainDefValidateVirtioDev() to retrieve the boolean
values of transitional and non-transitional models for each device,
is now being used inside qemuBuildVirtioDevStr() too.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>