qemu_hotplug: do not hotplug/hotunplug 'unassigned' hostdevs
The new address type 'unassigned' allows for a hostdev to be
managed by Libvirt, but not be assigned for guest usage. This
can allow us to execute PCI multifunction hotplug, with managed
mode, leaving non-zero functions unassigned.

For hotplug and unplug , we'll skip unassigned devices when
setting or evaluating aggregateSlotIdx and when we're about
to execute device_add/device_del in QEMU. For hotplug we'll
also skip adress assignment. For unplug we won't mark
the device for removal since we won't be executing device_del
on it.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>