qemu_domain.c: make qemuDomainGetMemoryModuleSizeAlignment() public
This function will be used in qemu_hotplug.c in the next patch.

While we're at it, we can remove the virDomainMemoryDef argument
since the alignment size doesn't change with this object. The
only existent caller can also be brought out of a loop which
iterates in a virDomainMemoryDef list.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>