virhostcpu.c: fix 'die' parsing for Power hosts
Commit 7b79ee2f78 makes assumptions about die_id parsing in
the sysfs that aren't true for Power hosts. In both Power8
and Power9 the die_id is set to -1:

$ cat /sys/devices/system/cpu/cpu0/topology/die_id
-1

This breaks the parsing being done in virHostCPUGetDie(),
which is using an uInt instead.

This isn't necessarily a PowerPC only behavior. Linux kernel commit
0e344d8c70 added in the former Documentation/cputopology.txt, now
Documentation/admin-guide/cputopology.rst, that:

---
To be consistent on all architectures, include/linux/topology.h
provides default definitions for any of the above macros that are
not defined by include/asm-XXX/topology.h:

1) topology_physical_package_id: -1
2) topology_die_id: -1
(...)
---

This means that it might be expected that an architecture that
does not implement the die_id element will mark it as -1 in
sysfs.

It is not necessary to revamp the die_id implementation because
of that. Instead, let's change the parsing of the die_id
in virHostCPUGetDie() to read an integer value and, in case
it's -1, default it to zero like in case of file not found.

Fixes: 7b79ee2f78bbf2af76df2f6466919e19ae05aeeb
Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>